Simplified Recoil System
Due to a switch to cinemachine virtual cameras over freelook camera (better collision detection) the recoil system needed to be changed to work with the new cameras. Recoil is now applied through the use of impulse sources. Unfortunately I had to remove firing patterns such as the S shape on assault rifle. Hoping to implement it back in later.